And that is part 2 done

+++ include/aoc.h
@@ -58,7 +58,7 @@ uint32_t day05_part1(void);
 uint64_t day05_part2(void);
 
 uint64_t day06_part1(void);
-uint32_t day06_part2(void);
+uint64_t day06_part2(void);
 
 #ifdef __cplusplus
 }
